+++ CHANGELOG.md
@@ -6,7 +6,11 @@ All notable changes to this project will be documented in this file.
 ### Added:
  - User *register* method
 ### Fixed: 
- - Pins *edit* method returns bool value
+ - Pins *edit* method returns bool value 
+### Changed:
+ - *login*, *isLoggedIn* and *logout* methods move to User provider
+ - Bot *login*, *isLoggedIn* and *logout* methods are deprecated
+ 
 
 ## [4.2.2] - 2016-06-19
 ### Fixed: